requirements.txt
google-generativeai
asyncio
psycopg2-binary
streamlit==1.51.0